requirements.txt
autopep8==1.6.0
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
cryptography==36.0.1
dnspython==2.1.0
filelock==3.4.0
Flask==2.0.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
future==0.18.2
gevent==21.12.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
install==1.3.5
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
pymongo==4.0.1
PySocks==1.7.1
python-dotenv==0.19.2
requests==2.26.0
requests-file==1.5.1
six==1.16.0
threaded==4.1.0
tldextract==3.1.2
toml==0.10.2
urllib3==1.26.7
Werkzeug==2.0.2
zope.event==4.5.0
zope.interface==5.4.0
